// src/modules/enviroment/application/exceptions/environment-response.enum.ts
export enum ENVIRONMENT_RESPONSE {
  ENVIRONMENT_NOT_FOUND_BY_COLLECTION_AND_USER = 'The user of the collection and the logged in user do not match',
  ENVIRONMENT_NOT_FOUND_BY_USER_ID = 'The user id does not match',
  ENVIRONMENT_COLLECTION_NOT_FOUND = 'The collection does not exist',
  ENVIRONMENT_FAILED_SAVE = 'Could not save variable, please try again',
  ENVIRONMENT_FAILED_DELETED = 'Could not deleted variable, please try again',
  ENVIRONMENT_FAILED_UPDATED = 'Could not updated variable, please try again',
}